Put the prescription form into another component
